stapdyn: Add to configure and Makefile
This is my rough attempt to bring stapdyn and dynsdt into the build.
Now configure can take --with-dyninst=DIR to set the location of Dyninst
headers and libraries.  This doesn't yet try to be arch-aware, only
x86_64.  It's also only looking for a build-tree layout, not distro-
installed dyninst files.

* Makefile.am: Add stapdyn and dynsdt to the build.
* configure: Process --with-dyninst, with a mild sanity check.
* README: Note DYNINSTAPI_RT_LIB and suggest LD_LIBRARY_PATH.
* TODO: Mark this task partially done.
* ...: Update all other autoreconf-generated files too.