loc2c-runtime.h: Add uread() and uwrite()
These are the userspace equivalents to kread() and kwrite().

For the (unused) STAPCONF_PROBE_KERNEL case, just define them as direct
equivalents.  Also add in missing kderef and store_kderef the same way.

For the architecture-implemented case, they are nearly identical to
kread and kwrite, just s/kderef/uderef/.  So to avoid duplication, I've
added in a layer of macro indirection to pass in the appropriate deref
function.  (Nevermind that the [ku]derefs are often identical -- oh,
what a tangled web we weave...)