runtime: Use typedef-typeof for safer function pointers
We use kallsyms_lookup_name to access several functions that are
publicly declared but not exported.  We were manually recreating the
call signatures for function pointers though, which is begging for
trouble.  This patch updates all those typedefs to use typeof on the
original function declaration, so we hopefully get compile-time breakage
if the call signature ever changes on us.