android: fix invalid install of staprun in output dir
Use 2 Android.mk's:
- One in src root for stap
- One in staprun for staprun & stapio

We do this to avoid a bug in the build system that triggers when some sources
are located in a directory that has the same name as the module to be built.
In this case staprun object output dir and staprun linked binary were confusing
the build system.

Signed-off-by: Dragos Tatulea <dragos.tatulea@intel.com>
Signed-off-by: Adrian Marius Negreanu <adrian.m.negreanu@intel.com>