PR6580: incremental unwinder with cache.
Implements a cached unwinder, allowing most backtrace tapset functions
to be implemented in terms of stack() without loss of performance.
stack() calls can be made repeatedly and in any order, and they will
use the results of a single unwind. (Works only with the DWARF
unwinder).

_stp_stack_kernel_print et. al. retain their prior behaviour,
including a number of fallbacks not available to the incremental
unwind. These fallbacks only emit backtrace strings, which can be
tokenized on the tapset end as a last resort.

* runtime/unwind/unwind.h -- define struct unwind_cache to store PCs
  obtained from unwinder.
* runtime/common_probe_context.h -- include two sets of unwinder
  context and cache, one for user side, one for kernel.
* tapsets.cxx -- probe prologue includes a small thing to mark
  the unwind caches as being in an uninitialized state.
* runtime/stack.c -- incremental unwinder implementation.
* runtime/stack-dwarf.c -- deleted. Code moved to stack.c since
  this is now the preferred unwind method.
* tapset/linux/[u]context-symbols.stp -- change stack(), ustack()
  to directly call incremental unwinder.