Fixed race condition in git_version.h generation
The stap and staprun targets were based on generation of a
git_version.h file. The generation is being done based on running the
git_version.sh script. However, there were two glitches in the code:
1. The generation script was being run with an output file that didn't
have the full path specified, using the current ($PWD) directory as
output file. And that's because the script itself was not "ready" to
be used with full path.
2. When building Android, all scripts are being run with current
directory, the topmost Android root directory. The git_version.sh
script was usnig the current directory to store the git_version.h,
just before copying it in the *intermediates* target directory. If two
or more instances of the git_version.sh script were being run, they
were using the very same git_version.h file from the Android tree
root. This leads to a race condition.

This patch solves both of these problems, even though it basically
addresses only the first one.

The git_version.sh script was modified to accept a full path when
specifying an output file. So, now, the script can be called directly
with the *intermediates* git_version.h file as parameter. Also, the
two Android.mk files were modified to call the git_version.sh script
properly.

Issue: AXIA-917
Change-Id: Ia89db4a48df90f059aa962866435892b5942e85f
Signed-off-by: Alin Dobre <alin.m.dobre@intel.com>