PR13486: Always output a frame_base if needed
Previously, loc2c only emitted code for frame_base if it was on the
first loc of the chain.  But in some cases, the first piece may not
reference the frame_base while later parts do.

In the cases found in the bug, the first piece was GNU_implicit_pointer,
which doesn't even emit code of its own.  But its target DIE did need
the frame_base to compute its value.  Since loc2c didn't realize that,
we ended up emitting code using frame_base without ever declaring it.

* loc2c.c (c_emit_location): Loop over the whole loc chain to determine
  if a frame_base is needed, and output the first one found.