hash: Include several subdirectories of the runtime
One advantage of having the runtime path in the hash is its mtime also
tells us when any files were modified (at least with most editors that
write a temp file then atomically rename).  For stap hackers like
myself, it would be nice to get this benefit for subdirectories of the
runtime too, so let's also add those to the hash.

* hash.cxx (get_base_hash): Add /transport, /unwind, and either /dyninst
  or /linux depending on the current runtime mode.
  (find_uprobes_hash): Update the uprobes paths since PR14179's move.