stapdyn: Use FD_CLOEXEC on _stp_mem_fd instead of O_CLOEXEC
O_CLOEXEC is only available since Linux 2.6.23, which is fairly old, but
we may still care to run on such systems.  Using fcntl FD_CLOEXEC can
accomplish the same thing, and we don't need to worry about the race of
other threads calling exec at the same time as our module load, because
the whole process will be frozen.