Set staprun's verbosity one less than stap
Prior to commit e520ea8, staprun would get one -v for s.verbosity>1 and
a second -v for s.verbosity>2.  That commit unbounded the number of -v
for staprun, but changed the off-by-one, and staprun and stapio are too
chatty for that.

This also now sets stapdyn verbosity the same way.

* buildrun.cxx (make_run_command): Give one less -v flag.
  (make_dyninst_run_command): Set stapdyn -v flags the same way.