stapdyn: Reduce LDADD to only those directly used
The only part of Dyninst that we actually use and need to link is
libdyninstAPI.  The rest of those libraries I had in dyninst_LIBS were
those needed indirectly by that one.  But trying to enumerate someone
else's dependencies is a recipe for failure -- e.g. it was already
broken by not including the new libdynElf too.

The reason I had this was to allow successful linking when using
--with-dyninst=/not/a/lib/searched/path without requiring also to set
LD_LIBRARY_PATH.  Since the enumeration is prone to break, and you'd
have to set LD_LIBRARY_PATH anyway to be able to run, I've decided it's
not worth trying to be fancy here.