stapdyn: Add tapset function print_regs
The underlying _stp_print_regs is more limited than the kernel version,
as some registers are not available, but this is still useful to see all
the information available to the probe.

Note: this makes it obvious that i686 actually doesn't have register
values right now.  Dyninst has documented that only POWER and AMD64
implement getRegisters, unfortunately.  This might be expanded by them
in a future release.

* runtime/regs.c: Split into linux/ and dyninst/ versions.
* tapset/context.stp (print_regs): Moved from linux/; now it's generic.
* stapdyn/mutatee.cxx (get_dwarf_registers): Add verbose logging, which
  also shows that i686 doesn't have registers. :(