PR11334: custom regex parser & generation wrapper
Further progress -- this one emits code that correctly returns 1 or 0
depending on whether the regex correctly matched or not.

TODO -- re2c seems to anchor the regex on the left of the string, so
to allow a match throughout the string we need to put .* (or similar)
on the front of the regex.

TODO -- .* and similar fail to match due to expecting '\n' to terminate
the string. This is an assumption of re2c which can be fixed.