PR11334: fix read-past-the-end error for .* and similar
Solved by the simple expedient of fixing the off-by-one error that caused
YYFILL to abort the match before the terminating '\0' could be returned,
and changing mkDot() (responsible for creating '.' regex) to forbid '\0'.

NOTE: This will have to be revisited due to reliance on explicit '\0'
matching. Namely, a tricky regex *might* try to break things by trying
to match '[s\0]*' or such. This will not read forbidden memory (YYFILL
is too airtight for that), but the match outcome will not be what one
would expect. So we should keep in mind that the regex syntax should
not allow '\0' as a regex match character.

* re2c-migrate/actions.cc (mkDot); detect '\0' instead of '\n' as string end.
* re2c-migrate/regcomp.cxx (emit_declaration): tweak YYFILL to include '\0'.
* re2c-migrate/regcomp.base/regtest.in.0: additional test case.