runtime: Remove dupes between map-gen and pmap-gen
Most of pmap-gen was a rehash of map-gen, just named pmap instead.  Only
a few of the pmap functions were actually unique to that situation, so
the rest have been terminated with extreme prejudice.

This also addresses a FIXME note I added in commit 57eedf945 about the
way map-gen and pmap-gen are included.  Made it in before the seventh
anniversary of that FIXME -- whew!

* translate.cxx (c_unparser::emit_map_type_instantiations): Set
  MAP_DO_PMAP to trigger the pmap-gen from within map-gen.
* runtime/map-gen.c (KEY_EQ_P): New, consolidate testing of all keys.
  (MAP_DO_PMAP): Use to include pmap-gen while the #defines are hot.
* runtime/pmap-gen.c: I love the smell of napalm in the morning...