stapdyn: Move alibi and timing stats to shm
In dyninst mode, the alibi and timing parts are now part of the main
session struct that lives in shared memory.  In kernel mode the same
parts are their own global arrays, but otherwise unchanged.

Also, the global stap_probes[] that used to hold these has no more
mutable members, so it is now const.

* runtime/dyninst/common_session_state.h: Add alibi and timing data to
  the session, and define functions to access them.
* runtime/linux/common_session_state.h: Move alibi and timing data to
  global arrays, and define the same access functions.
* tapsets.cxx (common_probe_entryfn_prologue): Use probe_alias() and
  probe_timing() to get the relevant data in runtime-neutral ways.
* translate.cxx (c_unparser::emit_module_exit): Ditto.
  (translate_pass): Define STP_PROBE_COUNT for sizing related arrays.
  Update the stap_probes[] initialization with a fixed index to access
  the relocated alias/timing data, and make the whole thing const.
* [everything else]: Always use const struct stap_probe pointers.