runtime: Updated support for uretprobes test kernels
This updates support for uretprobes kernels from here:
  http://repos.fedorapeople.org/repos/aarapov/kernel-uprobes/
... particularly kernel-3.7.0-99.uretprobes.fc18 built 13-Dec-2012.

The struct uretprobe_consumer is gone, and now both uprobes and
uretprobes use the same struct uprobe_consumer.  The signature of the
handler function has changed to include a long addr for the uretprobes'
true return address.  In the future that parameter will also be set for
plain uprobes to the equivalent of uprobe_get_swbp_addr().

I'm knowingly breaking compatibility with earlier uretprobes test
kernels.  As far as I know, those have never been officially shipped
anywhere, so that should be fine.  This does maintain compatibility with
the version of plain uprobes before the change in the handler signature.

* runtime/linux/uprobes-inode.c (struct stapiu_consumer): Collapse the
  uconsumer and uretconsumer back to a single entry.
  (stapiu_probe_prehandler): Fill in the given IP temporarily, then call
  the true handler from the translator.
  (stapiu_probe_prehandler_noaddr): Compatibility handler to grab the IP
  from uprobe_get_swbp_addr().
* runtime/linux/autoconf-inode-uprobes-noaddr.c: New autoconf to detect
  the former version of uprobes handler, without the addr parameter.
* buildrun.cxx (compile_pass): Use that autoconf.
* tapsets.cxx (uprobe_derived_probe_group::emit_module_inode_decls):
  Yield the IP-munging duties to uprobes-inode.c.