translate: for stap -t reports, identify probe index
For probe points that expand to multiple separate low-level kprobes,
e.g. for multiply-inlined functions, the stap -t report doesn't give
any hint as to which of those separate kprobes was invoked how many
times.  Normally, a user would not care, so that's fine, but sometimes
for debugging it's nice to match up the stap -t output to the stap -p3
stap_probes[] array elements.  So let's extend the -t output with an
undocumented index: NNN ditty.