examples: tweak deviceseeks.stp
Switch to using the ioblock_trace.request probe point (based on a
kernel tracepoint), which is more reliable than the
kprobe-on-inlined-function ioblock.request one, and makes more
information avilable.  (Unfortunately, this makes the script
incompatible with non-tracepoint kernels.)

Also, as per Jeff Moyer's correction, seeks are calculated as relative
to the end position of the prior transfer, not its origin.