PR11341: Add type safety to update_visitor::require()
This uses dynamic_cast<T*> to make a runtime check that the pointer that
was given to update_visitor::provide() is compatible with what the
caller required.  Such a cast can't be used on void*, so the ::stack is
now composed of 'visitable' pointers - a new base class for expressions
and statements.

In the process, 'indexable' is now also brought into the expression
hierarchy.  This looks a little odd since hist_op isn't really an
expression on its own, but otherwise this change untangles a lot of code
trying to deal with hist_op as an outlier, especially visit and print.

* staptree.h (update_visitor::require): Use dynamic_cast<> instead of a
  dangerous static_cast<>, and throw exceptions for bad cases.
  (struct visitable): New base class for expression and statement.
  (struct indexable): Use expression as a base class, and get rid of the
  now-unnecessary visit/print/tok middleman methods.
* staptree.cxx: Use direct indexable->visit/print/tok throughout.
* elaborate.cxx, parse.cxx: Ditto.