PR11341: Fix two improper update_visitor assignments
Both are cases where a too-specific require<target_symbol> was used,
inferred by the parameter type.  The result is immediately saved as a
generic expression pointer, so require<expression> is appropriate.

I believe this case is such that there was actually no harm caused here
by the formerly blithe static_cast, but it's still nice to see that such
improprieties are now detected.